+++ internal/interp/exec.go
@@ -102,6 +102,12 @@ func (in *Instance) runFrame(fn *binary.Func, locals *frame, st *stack, results,
 				}
 				continue
 			}
+			if ins.Prefix == 0xfd {
+				if err := in.execFD(ins, st); err != nil {
+					return err
+				}
+				continue
+			}
 			return unsupported(ins)
 		}
 		switch ins.Op {